move RoutingFunction to flink_translator
